feat: virtualize pokemon list
