+++ package.json
@@ -1,31 +1,33 @@
 {
-  "name": "paldex",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.33",
-    "@types/react-dom": "18.0.11",
-    "autoprefixer": "10.4.14",
-    "class-variance-authority": "^0.5.1",
-    "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "next": "13.3.0",
-    "postcss": "8.4.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "3.3.1",
-    "typescript": "5.0.4",
-    "zod": "^3.21.4"
-  },
-  "devDependencies": {
-    "prettier": "^2.8.7",
-    "prettier-plugin-tailwindcss": "^0.2.7"
-  }
+    "name": "paldex",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@types/node": "18.15.11",
+        "@types/react": "18.0.33",
+        "@types/react-dom": "18.0.11",
+        "autoprefixer": "10.4.14",
+        "class-variance-authority": "^0.5.1",
+        "eslint": "8.38.0",
+        "eslint-config-next": "13.3.0",
+        "next": "13.3.0",
+        "postcss": "8.4.21",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-virtuoso": "^4.2.0",
+        "sharp": "^0.32.0",
+        "tailwindcss": "3.3.1",
+        "typescript": "5.0.4",
+        "zod": "^3.21.4"
+    },
+    "devDependencies": {
+        "prettier": "^2.8.7",
+        "prettier-plugin-tailwindcss": "^0.2.7"
+    }
 }
