+++ src/app/layout.tsx
@@ -1,9 +1,12 @@
 import "./globals.css"
+import { Inter } from "next/font/google"
 
 export const metadata = {
     title: "Create Next App",
     description: "Generated by create next app",
 }
+
+const inter = Inter({ subsets: ["latin"] })
 
 export default function RootLayout({
     children,
@@ -12,7 +15,7 @@ export default function RootLayout({
 }) {
     return (
         <html lang="en">
-            <body>
+            <body className={inter.className}>
                 <main className="flex min-h-screen flex-col bg-[url(/wood-bg.jpeg)]">
                     {children}
                 </main>
